move go.mod and go.sum
